+++ sentry-android-core/src/main/java/io/sentry/android/core/AppState.java
@@ -8,6 +8,7 @@ import io.sentry.ILogger;
 import io.sentry.ISentryLifecycleToken;
 import io.sentry.NoOpLogger;
 import io.sentry.SentryLevel;
+import io.sentry.SentryOptions;
 import io.sentry.android.core.internal.util.AndroidThreadChecker;
 import io.sentry.util.AutoClosableReentrantLock;
 import java.io.Closeable;
@@ -35,9 +36,16 @@ public final class AppState implements Closeable {
 
   private volatile @Nullable Boolean inBackground = null;
 
+  @ApiStatus.Internal
   @TestOnly
-  void resetInstance() {
+  public void resetInstance() {
     instance = new AppState();
+  }
+
+  @ApiStatus.Internal
+  @TestOnly
+  public LifecycleObserver getLifecycleObserver() {
+    return lifecycleObserver;
   }
 
   public @Nullable Boolean isInBackground() {
@@ -48,7 +56,7 @@ public final class AppState implements Closeable {
     this.inBackground = inBackground;
   }
 
-  void addAppStateListener(final @NotNull AppStateListener listener) {
+  public void addAppStateListener(final @NotNull AppStateListener listener) {
     try (final @NotNull ISentryLifecycleToken ignored = lock.acquire()) {
       ensureLifecycleObserver(NoOpLogger.getInstance());
 
@@ -56,7 +64,7 @@ public final class AppState implements Closeable {
     }
   }
 
-  void removeAppStateListener(final @NotNull AppStateListener listener) {
+  public void removeAppStateListener(final @NotNull AppStateListener listener) {
     try (final @NotNull ISentryLifecycleToken ignored = lock.acquire()) {
       if (lifecycleObserver != null) {
         lifecycleObserver.listeners.remove(listener);
@@ -64,7 +72,8 @@ public final class AppState implements Closeable {
     }
   }
 
-  void registerLifecycleObserver(final @Nullable SentryAndroidOptions options) {
+  @ApiStatus.Internal
+  public void registerLifecycleObserver(final @Nullable SentryOptions options) {
     if (lifecycleObserver != null) {
       return;
     }
@@ -120,7 +129,8 @@ public final class AppState implements Closeable {
     }
   }
 
-  void unregisterLifecycleObserver() {
+  @ApiStatus.Internal
+  public void unregisterLifecycleObserver() {
     if (lifecycleObserver == null) {
       return;
     }
@@ -154,7 +164,8 @@ public final class AppState implements Closeable {
     unregisterLifecycleObserver();
   }
 
-  final class LifecycleObserver implements DefaultLifecycleObserver {
+  @ApiStatus.Internal
+  public final class LifecycleObserver implements DefaultLifecycleObserver {
     final List<AppStateListener> listeners =
         new CopyOnWriteArrayList<AppStateListener>() {
           @Override
@@ -170,12 +181,18 @@ public final class AppState implements Closeable {
           }
         };
 
+    @ApiStatus.Internal
+    @TestOnly
+    public List<AppStateListener> getListeners() {
+      return listeners;
+    }
+
     @Override
     public void onStart(@NonNull LifecycleOwner owner) {
+      setInBackground(false);
       for (AppStateListener listener : listeners) {
         listener.onForeground();
       }
-      setInBackground(false);
     }
 
     @Override