+++ sentry-android-core/src/main/java/io/sentry/android/core/ContextUtils.java
@@ -21,7 +21,6 @@ import io.sentry.SentryLevel;
 import io.sentry.SentryOptions;
 import io.sentry.android.core.util.AndroidLazyEvaluator;
 import io.sentry.protocol.App;
-import io.sentry.util.LazyEvaluator;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileReader;
@@ -90,20 +89,6 @@ public final class ContextUtils {
 
   // to avoid doing a bunch of Binder calls we use LazyEvaluator to cache the values that are static
   // during the app process running
-
-  private static final @NotNull LazyEvaluator<Boolean> isForegroundImportance =
-      new LazyEvaluator<>(
-          () -> {
-            try {
-              final ActivityManager.RunningAppProcessInfo appProcessInfo =
-                  new ActivityManager.RunningAppProcessInfo();
-              ActivityManager.getMyMemoryState(appProcessInfo);
-              return appProcessInfo.importance == IMPORTANCE_FOREGROUND;
-            } catch (Throwable ignored) {
-              // should never happen
-            }
-            return false;
-          });
 
   /**
    * Since this packageInfo uses flags 0 we can assume it's static and cache it as the package name
@@ -284,7 +269,15 @@ public final class ContextUtils {
    */
   @ApiStatus.Internal
   public static boolean isForegroundImportance() {
-    return isForegroundImportance.getValue();
+    try {
+      final ActivityManager.RunningAppProcessInfo appProcessInfo =
+          new ActivityManager.RunningAppProcessInfo();
+      ActivityManager.getMyMemoryState(appProcessInfo);
+      return appProcessInfo.importance == IMPORTANCE_FOREGROUND;
+    } catch (Throwable ignored) {
+      // should never happen
+    }
+    return false;
   }
 
   /**
@@ -544,7 +537,6 @@ public final class ContextUtils {
 
   @TestOnly
   static void resetInstance() {
-    isForegroundImportance.resetValue();
     staticPackageInfo33.resetValue();
     staticPackageInfo.resetValue();
     applicationName.resetValue();