+++ sentry-samples/sentry-samples-android/sdkperf/screen_flap.sh
@@ -1,0 +1,15 @@
+#!/bin/bash
+
+# Loop 4 times to toggle Screen
+for i in {1..4}
+do
+  echo "[$i] Turning screen off..."
+  adb shell input keyevent 223
+  sleep 5
+
+  echo "[$i] Turning screen on..."
+  adb shell input keyevent 224
+  sleep 5
+done
+
+echo "Done flapping Screen 4 times."